:sparkles: add prototype for retrieving shared lcs assets
Adds internal function cmdlets for the unofficial LCS API endpoints GetSharedAssets (Get-LCSSharedAssetFile) and GetFileAsset (Get-LSFileAsset).

Adds cmdlet Get-D365LCSSharedAssetFile to retrieve a list of shared asset files. The list is first retrieved with Get-LCSSharedAssetFile. After filtering the list, the id of each entry is used with Get-LCSFileAsset to retrieve the file location and add it to the result.

Adds cmdlet Get-D365LCSFileAsset to test retrieval of a shared asset file by id directly.