Also search %PATH% for sqlpackage.exe
The installation documentation for SqlPackage says:
Installing SqlPackage as a global tool makes it available on your path
as sqlpackage and is the recommended method to install SqlPackage for
Windows, macOS, and Linux.

In line with Microsoft's expectations, also search %PATH% if
sqlpackage is not found in the d365fo.tools default path
($Script:SqlPackagePath). If it is found in the %PATH% then update the
default path with the actual path to sqlpackage.exe.

Author: Colin Daley
Date: 2024-03-24