update default SqlPackage version
This updates the default version of SqlPackage that d365fo.tools installs to 162.2.111.2 (the first .NET 8 version).

In addition, it replaces the -SkipExtractFromPage parameter and the corresponding web scraping logic with the -Latest parameter and the evergreen link https://aka.ms/sqlpackage-windows

Note that unlike the behavior from when the web scraping logic was still working, the latest version of SqlPackage is NOT installed by default. Instead, the fixed version 162.2.111.2 is installed (which may be updated to later versions in the future). Reasons for this change in behavior are discussed in #747.

fixes #747